ScrapItems components to a new file
